C++: Improve rwlock
The yield function seems to be a system_call which creates more latency
and is only useful when the # of threads > # of CPUs.
Ideally we may use the _mm_pause intrinsics but obviously it is not in
the C++ standard.
The current code only calls yield after some attempts to get the lock
failed. At the moment the number is hard-coded to 16.