C++: Add contiguousmap
This is a binary tree not based on pointers and nodes, but in a linear
buffer and indices.
This is interesting from the point of view of avoid allocations and
recursions at a cost of a more expensive rebalance.
The extra cost of reallocations when extra space is needed, is also
important to consider.